server/apps/users/serializers.py
from rest_framework import serializers
from server.apps.users.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email', 'created_at', 'is_active')
